Time: 0 ms (100%), Space: 8 MB (99.78%) - LeetHub

+++ 1797-goal-parser-interpretation/1797-goal-parser-interpretation.cpp
@@ -1,11 +1,11 @@
 class Solution {
 public:
     string interpret(string command) {
-        size_t commandLen = command.length();
+        int commandLen = command.length();
         string answer;
         answer.reserve(commandLen);
 
-        for (size_t i = 0; i < commandLen; ++i)
+        for (int i = 0; i < commandLen; ++i)
         {
             char ch = command[i];
